Add files via upload
In this project, my aim is solve a maze with certain rules. I have 3 classes which are 
MazeSolution , Node and Stack. To be able to reach that goal , I used different implementation I 
tried the conditions given to us in the question in the if -else structure by taking the input from the 
user.In node class I create x and y variables which I think these are x and y axes and a player moved 
there. And also In Stack class I determine push and pop methods as we know pusmh methods help us
to insert a element in stack , and pop methods removes the element in the stack .
“Walls are represented with “+”, “-“, and“|” characters we can’t move forward when we saw this 
characters .”
“.Available paths are represented with lower case characters. When we see these characters, you 
can move forward.”
“Treasures are represented with capital “E” characters. When we see this character, this means 
that you can find a path successfully.”